feat: split tests
